added changes to restructure the events folder
