--- README.md
@@ -1,38 +1,0 @@
-#Git Flow/Rules
-## ****DO NOT PULL FROM/CLONE FROM MASTER, DEVELOPMENT ONLY*****
-**Steps:** 
- 1. git clone from development branch
- 
- 2. create a branch with the feature name you are working on and your initials ex: nav-bar-tb
- 
- 3. check trello/slack to see if anyone is working on a feature that interacts or is dependent on the feature you are working on. Coordinate/peer program with them to find out the best approach
- 
- 4. commit frequently 
- 
- 5. any methods or logic that is longer than 3-5 lines comment what it does to make it easier/faster for anyone coming in behind you to figure out what you have done
- 
- 6. when you feel ready to do a pull request.. **DON'T!** Pull from development first, check the incoming changes and **VERIFY** that everything still works BEFORE you do your pull request!
- 
- 7. any pull requests ***MUST HAVE A CODE REVIEW*** from the Git Master for the week. Just IM them and let them know to look and approve. ***DONT JUST APPROVE THE CODE, ACTUALLY REVIEW IT*** find at least one thing they did well and one thing they can improve (if possible) this will be good practice for when we have to do it in the real world. If there are some changes you recommend try to zoom with them and go over why you think this would work better/ what your thought process is with it. I know it takes time, but everyone will come out a better dev by the end of our awesome project if we do this. 
- 
-8. once your pull request is approved to go to development, whoever approves and pull requests needs to let the whole team know that there have been changes and specify what feature was changed/add so people know to expect changes regarding what they are working on.
-
-9. After the merge was successful ***delete that branch***. 
-
-10. whoever submitted the branch will create a new branch with the next features name they will be working on
-
-11. We will merge to master at the end of each sprint. All PR's must be approved and merged beforehand!
-
-
-##GIT MASTER SCHEDULE
-***Git Masters, if you need a PR merged ask the person who was GM before you***
-
-Week 2: Taylor Belk
-Week 3: Noah Stoeckel
-Week 4: Arpita Sinha
-Week 5: Andrey Yurlov
-Week 6: Alex Bezruchko
-Week 7: Taylor Belk
-Week 8: Noah Stoeckel
-Week 9: Arpita Sinha
-Week 10: Alex Bezruchko
