+++ pizza-time/README.md
@@ -1,68 +1,38 @@
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+#Git Flow/Rules
+## ****DO NOT PULL FROM/CLONE FROM MASTER, DEVELOPMENT ONLY*****
+**Steps:** 
+ 1. git clone from development branch
+ 
+ 2. create a branch with the feature name you are working on and your initials ex: nav-bar-tb
+ 
+ 3. check trello/slack to see if anyone is working on a feature that interacts or is dependent on the feature you are working on. Coordinate/peer program with them to find out the best approach
+ 
+ 4. commit frequently 
+ 
+ 5. any methods or logic that is longer than 3-5 lines comment what it does to make it easier/faster for anyone coming in behind you to figure out what you have done
+ 
+ 6. when you feel ready to do a pull request.. **DON'T!** Pull from development first, check the incoming changes and **VERIFY** that everything still works BEFORE you do your pull request!
+ 
+ 7. any pull requests ***MUST HAVE A CODE REVIEW*** from the Git Master for the week. Just IM them and let them know to look and approve. ***DONT JUST APPROVE THE CODE, ACTUALLY REVIEW IT*** find at least one thing they did well and one thing they can improve (if possible) this will be good practice for when we have to do it in the real world. If there are some changes you recommend try to zoom with them and go over why you think this would work better/ what your thought process is with it. I know it takes time, but everyone will come out a better dev by the end of our awesome project if we do this. 
+ 
+8. once your pull request is approved to go to development, whoever approves and pull requests needs to let the whole team know that there have been changes and specify what feature was changed/add so people know to expect changes regarding what they are working on.
 
-## Available Scripts
+9. After the merge was successful ***delete that branch***. 
 
-In the project directory, you can run:
+10. whoever submitted the branch will create a new branch with the next features name they will be working on
 
-### `npm start`
+11. We will merge to master at the end of each sprint. All PR's must be approved and merged beforehand!
 
-Runs the app in the development mode.<br>
-Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
 
-The page will reload if you make edits.<br>
-You will also see any lint errors in the console.
+##GIT MASTER SCHEDULE
+***Git Masters, if you need a PR merged ask the person who was GM before you***
 
-### `npm test`
-
-Launches the test runner in the interactive watch mode.<br>
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
-
-### `npm run build`
-
-Builds the app for production to the `build` folder.<br>
-It correctly bundles React in production mode and optimizes the build for the best performance.
-
-The build is minified and the filenames include the hashes.<br>
-Your app is ready to be deployed!
-
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
-
-### `npm run eject`
-
-**Note: this is a one-way operation. Once you `eject`, you can’t go back!**
-
-If you aren’t satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
-
-Instead, it will copy all the configuration files and the transitive dependencies (Webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you’re on your own.
-
-You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
-
-## Learn More
-
-You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
-
-To learn React, check out the [React documentation](https://reactjs.org/).
-
-### Code Splitting
-
-This section has moved here: https://facebook.github.io/create-react-app/docs/code-splitting
-
-### Analyzing the Bundle Size
-
-This section has moved here: https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size
-
-### Making a Progressive Web App
-
-This section has moved here: https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app
-
-### Advanced Configuration
-
-This section has moved here: https://facebook.github.io/create-react-app/docs/advanced-configuration
-
-### Deployment
-
-This section has moved here: https://facebook.github.io/create-react-app/docs/deployment
-
-### `npm run build` fails to minify
-
-This section has moved here: https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify
+Week 2: Taylor Belk
+Week 3: Noah Stoeckel
+Week 4: Arpita Sinha
+Week 5: Andrey Yurlov
+Week 6: Alex Bezruchko
+Week 7: Taylor Belk
+Week 8: Noah Stoeckel
+Week 9: Arpita Sinha
+Week 10: Alex Bezruchko
