+++ pizza-time/src/components/events/create-new-event/friendPicker/friendPicker.js
