+++ pizza-time/src/components/events/create-new-event/search/list/places-list.js
